+++ website/src/stores/solana.ts
@@ -1,8 +1,8 @@
 import {defineStore} from 'pinia';
 import {clusterApiUrl, Commitment, Connection} from '@solana/web3.js';
-import {Router} from 'src/router';
 import {validateUrl} from 'src/utils/urls';
-import {COMMITMENT_SETTINGS_KEY, NETWORK_SETTINGS_KEY} from 'src/constants';
+import {COMMITMENT_SETTINGS_KEY, NETWORK_SETTINGS_KEY, WALLET_AUTO_CONNECT} from 'src/constants';
+import {useRouterStore} from 'stores/router';
 
 const networks: Record<string, string> = {
     'mainnet-beta': clusterApiUrl('mainnet-beta'),
@@ -15,12 +15,15 @@ export const useSolanaStore = defineStore('solana', {
     state: () => ({
         network: clusterApiUrl('mainnet-beta'),
         commitment: 'confirmed' as Commitment,
+        walletAutoConnect: localStorage.getItem(WALLET_AUTO_CONNECT) === 'true',
     }),
     getters: {
         connection: state => new Connection(state.network, state.commitment),
     },
     actions: {
         async setNetwork(network: string) {
+            const routerStore = useRouterStore();
+
             for (const net in networks) {
                 const url = networks[net];
 
@@ -33,13 +36,7 @@ export const useSolanaStore = defineStore('solana', {
                         localStorage.setItem(NETWORK_SETTINGS_KEY, net);
                     }
 
-                    await Router.replace({
-                        query: {
-                            ...Router.currentRoute.value.query,
-                            network: net === 'mainnet-beta' ? undefined : net,
-                        },
-                    });
-
+                    await routerStore.setQueryEntry(NETWORK_SETTINGS_KEY, net === 'mainnet-beta' ? undefined : net);
                     return;
                 }
             }
@@ -48,39 +45,35 @@ export const useSolanaStore = defineStore('solana', {
                 this.network = network;
                 localStorage.setItem(NETWORK_SETTINGS_KEY, network);
 
-                await Router.replace({
-                    query: {
-                        ...Router.currentRoute.value.query,
-                        network,
-                    },
-                });
+                await routerStore.setQueryEntry(NETWORK_SETTINGS_KEY, network);
             }
         },
         async setCommitment(commitment: string) {
+            const routerStore = useRouterStore();
+
             switch (commitment) {
                 case 'confirmed':
                     this.commitment = commitment;
                     localStorage.removeItem(COMMITMENT_SETTINGS_KEY);
 
-                    await Router.replace({
-                        query: {
-                            ...Router.currentRoute.value.query,
-                            commitment: undefined,
-                        },
-                    });
+                    await routerStore.setQueryEntry(COMMITMENT_SETTINGS_KEY, undefined);
                     break;
                 case 'processed':
                 case 'finalized':
                     this.commitment = commitment;
                     localStorage.setItem(COMMITMENT_SETTINGS_KEY, commitment);
 
-                    await Router.replace({
-                        query: {
-                            ...Router.currentRoute.value.query,
-                            commitment,
-                        },
-                    });
+                    await routerStore.setQueryEntry(COMMITMENT_SETTINGS_KEY, commitment);
                     break;
+            }
+        },
+        async setWalletAutoConnect(autoConnect: boolean) {
+            this.walletAutoConnect = autoConnect;
+
+            if (autoConnect) {
+                localStorage.setItem(WALLET_AUTO_CONNECT, autoConnect.toString());
+            } else {
+                localStorage.removeItem(WALLET_AUTO_CONNECT);
             }
         },
     },